Add a stride scheduler

+++ proc.h
@@ -51,7 +51,7 @@ struct proc {
   char name[16];               // Process name (debugging)
   int syscall_counter;         // A counter to count the number of system calls the process has done so far.
   int tickets;                 // Tickets
-  float stride;                //Stride
+  int stride;                  //Stride
 };
 
 // Process memory is laid out contiguously, low addresses first:
